# Conflicts:
#	pinot-common/src/main/java/org/apache/pinot/common/metrics/ControllerMeter.java
#	pinot-common/src/main/java/org/apache/pinot/common/utils/config/TableConfigSerDeUtils.java
#	pinot-controller/src/main/java/org/apache/pinot/controller/helix/core/PinotHelixResourceManager.java
#	pinot-integration-tests/src/test/java/org/apache/pinot/integration/tests/OfflineClusterIntegrationTest.java
#	pinot-segment-local/src/test/java/org/apache/pinot/segment/local/utils/TableConfigUtilsTest.java
#	pinot-server/src/main/java/org/apache/pinot/server/starter/ServerInstance.java
#	pinot-server/src/main/java/org/apache/pinot/server/starter/helix/BaseServerStarter.java
#	pinot-spi/src/main/java/org/apache/pinot/spi/config/table/TableConfig.java
#	pinot-spi/src/main/java/org/apache/pinot/spi/utils/builder/TableConfigBuilder.java

+++ pinot-plugins/pinot-file-system/pinot-hdfs/src/main/java/org/apache/pinot/plugin/filesystem/HadoopPinotFS.java
@@ -81,11 +81,17 @@ public class HadoopPinotFS extends BasePinotFS {
   @Override
   public boolean delete(URI segmentUri, boolean forceDelete)
       throws IOException {
-    // Returns false if we are moving a directory and that directory is not empty
+    Path path = new Path(segmentUri);
+
+    if (!_hadoopFS.exists(path)) {
+      return true;
+    }
+
     if (isDirectory(segmentUri) && listFiles(segmentUri, false).length > 0 && !forceDelete) {
       return false;
     }
-    return _hadoopFS.delete(new Path(segmentUri), true);
+
+    return _hadoopFS.delete(path, true);
   }
 
   @Override
@@ -177,6 +183,9 @@ public class HadoopPinotFS extends BasePinotFS {
       throws IOException {
     // _hadoopFS.listFiles(path, false) will not return directories as files, thus use listStatus(path) here.
     FileStatus[] files = _hadoopFS.listStatus(path);
+    if (files == null) {
+      throw new IOException("FileSystem.listStatus() returned null for path: " + path);
+    }
     for (FileStatus file : files) {
       visitor.accept(file);
       if (file.isDirectory() && recursive) {
@@ -198,7 +207,7 @@ public class HadoopPinotFS extends BasePinotFS {
         throw new RuntimeException("_hadoopFS client is not initialized when trying to copy files");
       }
       if (_hadoopFS.isDirectory(remoteFile)) {
-        throw new IllegalArgumentException(srcUri.toString() + " is a direactory");
+        throw new IllegalArgumentException(srcUri.toString() + " is a directory");
       }
       long startMs = System.currentTimeMillis();
       _hadoopFS.copyToLocalFile(remoteFile, localFile);
@@ -228,23 +237,19 @@ public class HadoopPinotFS extends BasePinotFS {
   }
 
   @Override
-  public boolean isDirectory(URI uri) {
+  public boolean isDirectory(URI uri)
+      throws IOException {
     try {
       return _hadoopFS.getFileStatus(new Path(uri)).isDirectory();
-    } catch (IOException e) {
-      LOGGER.error("Could not get file status for {}", uri, e);
-      throw new RuntimeException(e);
+    } catch (FileNotFoundException e) {
+      return false;
     }
   }
 
   @Override
-  public long lastModified(URI uri) {
-    try {
-      return _hadoopFS.getFileStatus(new Path(uri)).getModificationTime();
-    } catch (IOException e) {
-      LOGGER.error("Could not get file status for {}", uri, e);
-      throw new RuntimeException(e);
-    }
+  public long lastModified(URI uri)
+      throws IOException {
+    return _hadoopFS.getFileStatus(new Path(uri)).getModificationTime();
   }
 
   @Override
@@ -252,8 +257,9 @@ public class HadoopPinotFS extends BasePinotFS {
       throws IOException {
     Path path = new Path(uri);
     if (!exists(uri)) {
-      FSDataOutputStream fos = _hadoopFS.create(path);
-      fos.close();
+      try (FSDataOutputStream fos = _hadoopFS.create(path)) {
+        // create an empty file; stream closed by try-with-resources
+      }
     } else {
       _hadoopFS.setTimes(path, System.currentTimeMillis(), -1);
     }
@@ -301,7 +307,9 @@ public class HadoopPinotFS extends BasePinotFS {
   @Override
   public void close()
       throws IOException {
-    _hadoopFS.close();
+    if (_hadoopFS != null) {
+      _hadoopFS.close();
+    }
     super.close();
   }
 }
